% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hashmap.R
\name{delete.hashmap}
\alias{delete.hashmap}
\title{Remove elements from hashmap}
\usage{
\method{delete}{hashmap}(where, what)
}
\arguments{
\item{where}{hashmap from where you want to remove keys and values}

\item{what}{deleted stuff}
}
\description{
Remove elements from hashmap
}
