% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hashmap.R
\name{has.key}
\alias{has.key}
\title{Test if hashmap has a key}
\usage{
has.key(where, key)
}
\arguments{
\item{where}{a hashmap}

\item{key}{what key, or keys, can be a list or vector}
}
\description{
Test if hashmap has a key
}
