% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{internal-functions}
\alias{internal-functions}
\alias{.left_outer_join_impl}
\alias{.left_outer_join_impl}
\alias{.right_outer_join_impl}
\alias{.inner_join_impl}
\alias{.full_outer_join_impl}
\title{Hashmap internal functions}
\usage{
.left_outer_join_impl(x, y)

.right_outer_join_impl(x, y)

.inner_join_impl(x, y)

.full_outer_join_impl(x, y)
}
\arguments{
\item{x}{an external pointer to a \code{HashMap}}

\item{y}{an external pointer to a \code{HashMap}}
}
\description{
Hashmap internal functions
}
\details{
These functions are intended for internal use only; do not
  call them directly.
}
