% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\name{plot}
\alias{plot}
\alias{plot.Rcpp_Hashmap}
\title{Plot method for Hashmap class}
\usage{
\method{plot}{Rcpp_Hashmap}(x, ...)
}
\arguments{
\item{x}{an object created by a call to \code{hashmap}}

\item{\dots}{arguments passed to \code{plot}}
}
\description{
Plot method for Hashmap class
}
\details{
A convenience function which simply calls \code{plot} using
 \code{x$keys()} and \code{x$values()} as plotting coordinates.
}
\examples{
x <- hashmap(1:20, rnorm(20))
plot(x)
plot(x, type = 'p', pch = 20, col = 'red')

y <- hashmap(Sys.Date() + 1:20, rnorm(20))
plot(y, type = 'h', col = 'blue', lwd = 3)
}

