% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{plugin}
\alias{plugin}
\title{Rcpp Plugin}
\usage{
plugin
}
\description{
An Rcpp plugin setting the correct linker flags for
 Rcpp files using the C++ API of hashmap.
}
\details{
In order to access the C++ API of hashmap, users must add
 \code{// [[Rcpp::plugins(hashmap)]]} and \code{#include <hashmap.h>}
 to the top of their Rcpp source files.
}

