% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hashmap.R
\name{rm.hashmap}
\alias{rm.hashmap}
\title{Remove elements from hashmap}
\usage{
rm.hashmap(where, ...)
}
\arguments{
\item{where}{hashmap from where you want to remove keys and values}

\item{...}{deleted stuff}
}
\description{
Remove elements from hashmap
}
