% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hashmap.R
\name{add.hashmap}
\alias{add.hashmap}
\title{Add values to hashmap}
\usage{
\method{add}{hashmap}(where, ...)
}
\arguments{
\item{where}{hashmap where you want to add keys and values}

\item{...}{Added stuff}
}
\description{
Add values to hashmap
}
