% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hashmap.R
\name{hashmap}
\alias{Rcpp_hashmap}
\alias{hashmap}
\title{The faster hashmap}
\usage{
new(hashmap, x, y)
}
\arguments{
\item{x}{an atomic vector of type \code{integer},
     \code{numeric}, or \code{character} representing
     the lookup keys}

\item{y}{an atomic vector of type \code{integer},
     \code{numeric}, or \code{character} representing
     the corresponding lookup values}
}
\value{
a \code{hashmap} object
}
\description{
An S4 class for fast key-value storage
     of atomic vectors
}
\details{
The faster hashmap
}
\examples{
x <- replicate(10e3,
    paste0(sample(letters, 12, TRUE),
           collapse = "")
)
y <- rnorm(length(x))
z <- sample(x, 100)

H <- new(hashmap, x, y)

all.equal(y[match(z, x)], H$find_values(z))

\dontrun{
microbenchmark::microbenchmark(
    "R" = y[match(z, x)],
    "H" = H$find_values(z),
    times = 500L
)
}
}

