% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hashmap.R
\name{hashmap}
\alias{hashmap}
\title{The faster hashmap}
\usage{
hashmap(keys, values, ...)
}
\arguments{
\item{keys}{an atomic vector of type \code{integer},
     \code{numeric}, or \code{character} representing
     the lookup keys}

\item{values}{an atomic vector of type \code{integer},
     \code{numeric}, \code{character}, or \code{logical}
     representing the corresponding lookup values}

\item{...}{other arguments passed to \code{new} when constructing
     the instance}
}
\value{
a \code{hashmap} object
}
\description{
An S4 class for fast key-value storage
     of atomic vectors
}
\details{
The faster hashmap
}
\examples{
x <- replicate(10e3,
    paste0(sample(letters, 12, TRUE),
           collapse = "")
)
y <- rnorm(length(x))
z <- sample(x, 100)

H <- hashmap(x, y)

all.equal(y[match(z, x)], H[[z]])

\dontrun{
microbenchmark::microbenchmark(
    "R" = y[match(z, x)],
    "H" = H[[z]],
    times = 500L
)
}
}

