% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hashmap.R
\name{hashmap}
\alias{hashmap}
\title{Create the hashmap class}
\usage{
hashmap(initial = NULL)
}
\arguments{
\item{initial}{Optional list of keys and values to be included in the hashmap}
}
\description{
Create the hashmap class
}
